refactor: Define funciones auxiliares en un modulo separado
